+++ examples/storybook/stories/Ricos/RicosPublishStory.tsx
@@ -6,7 +6,10 @@ import { pluginGallery } from 'wix-rich-content-plugin-gallery';
 import { pluginPoll } from 'wix-rich-content-plugin-social-polls';
 import MobileDetect from 'mobile-detect';
 import ActionButton from '../Components/ActionButton';
-import { withEditorContext, EditorEventsProps } from 'wix-rich-content-editor-common';
+import {
+  withEditorContext,
+  EditorEventsProps,
+} from 'wix-rich-content-editor-common/libs/EditorEventsContext';
 
 const mobileDetect = new MobileDetect(window.navigator.userAgent);
 const plugins = [pluginImage(), pluginGallery(), pluginPoll()];