+++ packages/plugin-social-polls/web/src/components/PollEditor.jsx
@@ -1,11 +1,10 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
+import { isPluginFocused } from 'wix-rich-content-editor-common';
 import {
   WithEditorEventsProps,
-  EditorEventsContext as EditorEventsContext1,
-  isPluginFocused,
-} from 'wix-rich-content-editor-common';
-import { EditorEventsContext as EditorEventsContext2 } from 'wix-rich-content-editor-common/libs/EditorEventsContext';
+  withEditorEvents,
+} from 'wix-rich-content-editor-common/libs/EditorEventsContext';
 
 import { Poll } from './Poll';
 import { PollContextProvider } from './poll-context';
@@ -64,8 +63,7 @@ class PollEditorComponent extends PureComponent {
       t,
       theme,
       isMobile,
-      editorEvents1,
-      editorEvents2,
+      editorEvents,
     } = this.props;
 
     return (
@@ -86,16 +84,7 @@ class PollEditorComponent extends PureComponent {
         }}
       >
         <PollContextProvider
-          editorEvents={{
-            subscribe: (event, cb) => {
-              editorEvents1?.subscribe(event, cb);
-              editorEvents2?.subscribe(event, cb);
-            },
-            unsubscribe: (event, cb) => {
-              editorEvents1?.unsubscribe(event, cb);
-              editorEvents2?.unsubscribe(event, cb);
-            },
-          }}
+          editorEvents={editorEvents}
           settings={settings}
           poll={componentData.poll}
           setPoll={this.setPoll}
@@ -108,18 +97,4 @@ class PollEditorComponent extends PureComponent {
   }
 }
 
-export const PollEditor = props => (
-  <EditorEventsContext1.Consumer>
-    {contextValue1 => (
-      <EditorEventsContext2.Consumer>
-        {contextValue2 => (
-          <PollEditorComponent
-            editorEvents1={contextValue1}
-            editorEvents2={contextValue2}
-            {...props}
-          />
-        )}
-      </EditorEventsContext2.Consumer>
-    )}
-  </EditorEventsContext1.Consumer>
-);
+export const PollEditor = withEditorEvents(PollEditorComponent);