+++ packages/ricos-common/web/src/RicosTypes.ts
@@ -57,9 +57,7 @@ export interface RicosEditorProps extends RicosProps {
   onBusyChange?: OnBusyChangeFunction;
   injectedContent?: DraftContent;
   maxTextLength?: number;
-  editorEvents1?: EditorEvents;
-  editorEvents2?: EditorEvents;
-
+  editorEvents?: EditorEvents;
   /* Changes to this interface should also be reflected in the API docs */
 }